+++ packages/core/src/render3/instructions/all.ts
@@ -27,3 +27,4 @@
  */
 export * from './instructions';
 export * from './styling_instructions';
+export * from './property';
